Update imports to use 'dist/types'
