+++ ui-packages/packages/components-common/src/types/types.ts
@@ -23,3 +23,30 @@ export enum SCHEMA_VERSION {
 export interface FormRendererApi {
   doReset: () => void;
 }
+
+export enum FormType {
+  HTML = 'HTML',
+  TSX = 'TSX'
+}
+
+export type FormInfo = {
+  type: FormType;
+  name: string;
+  lastModified: Date;
+};
+
+export type FormConfiguration = {
+  resources: FormResources;
+  schema: string;
+};
+
+export type FormResources = {
+  scripts: Record<string, string>;
+  styles: Record<string, string>;
+};
+
+export type CustomForm = {
+  formInfo: FormInfo;
+  source: string;
+  configuration: FormConfiguration;
+};
