+++ ui-packages/packages/process-form/src/index.ts
@@ -19,4 +19,3 @@
 export * from './api';
 export * from './embedded';
 export * from './envelope';
-export * from './types';
