+++ ui-packages/packages/task-form/src/embedded/EmbeddedTaskFormChannelApiImpl.ts
@@ -17,7 +17,7 @@
  * under the License.
  */
 import { TaskFormChannelApi, TaskFormDriver } from '../api';
-import { CustomForm } from '../types';
+import { CustomForm } from '@kogito-apps/components-common/dist/types';
 
 export class EmbeddedTaskFormChannelApiImpl implements TaskFormChannelApi {
   constructor(private readonly driver: TaskFormDriver) {}
