+++ ui-packages/packages/task-form/src/envelope/TaskFormEnvelopeView.tsx
@@ -35,7 +35,7 @@ import {
   KogitoEmptyState,
   KogitoEmptyStateType
 } from '@kogito-apps/components-common/dist/components/KogitoEmptyState';
-import { CustomForm } from '../types';
+import { CustomForm } from '@kogito-apps/components-common/dist/types';
 import { TaskFormEnvelopeViewDriver } from './TaskFormEnvelopeViewDriver';
 import CustomTaskFormDisplayer from './components/CustomTaskFormDisplayer/CustomTaskFormDisplayer';
 import TaskForm from './components/TaskForm/TaskForm';
